requirements.txt
apache-beam[gcp]
google-cloud
google-cloud-bigquery==1.17.0
google-cloud-dataflow==2.4.0
google-cloud-datastore==1.8.0
google-cloud-pubsub==0.42.1
google-cloud-storage==1.17.0
google-cloud-vision==0.38.0
mock==2.0.0
numpy==1.14.2
six==1.12.0

